Criação de serviço para exclusão de usuários
Closes #43

+++ back/pcvt/src/main/java/br/ufpe/cin/pcvt/controllers/UserController.java
@@ -35,7 +35,11 @@ public class UserController {
 		return repository.get(key);
 	}
 
-	public void remove(Integer key) {
+	public void remove(Integer key) throws UserNotFoundException {
+		User user = repository.get(key);
+		if(user == null)
+			throw new UserNotFoundException(key.toString());
+
 		repository.remove(key);
 	}
